Initial asan cleanups
This patch defines a new asan_shadow_offset target macro, instead of
having a mere macro in the asan.c file.  It becomes thus cleaner to
define the target macro for targets that supports asan, namely x86 for
now.  The ASAN_SHADOW_SHIFT (which, along with the asan_shadow_offset
constant, is used to compute the address of the shadow memory byte for
a given memory address) is defined in asan.h.

2013-06-03  Manjunath Matti <mmatti@caviumnetworks.com>

        Backported libsanitizer support.

        2012-11-12  Jakub Jelinek  <jakub@redhat.com>

        * toplev.c (process_options): Warn and turn off
        -faddress-sanitizer if not supported by target.
        * asan.c: Include target.h.
        (asan_scale, asan_offset_log_32, asan_offset_log_64,
        asan_offset_log): Removed.
        (build_check_stmt): Use ASAN_SHADOW_SHIFT and
        targetm.asan_shadow_offset ().
        (asan_instrument): Don't initialize asan_offset_log.
        * asan.h (ASAN_SHADOW_SHIFT): Define.
        * target.def (TARGET_ASAN_SHADOW_OFFSET): New hook.
        * doc/tm.texi.in (TARGET_ASAN_SHADOW_OFFSET): Add it.
        * doc/tm.texi: Regenerated.
        * Makefile.in (asan.o): Depend on $(TARGET_H).
        * config/i386/i386.c (ix86_asan_shadow_offset): New function.
        (TARGET_ASAN_SHADOW_OFFSET): Define.

git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@193433 138bc75d-0d04-0410-961f-82ee72b054a4